Implement pip version range
This patch implements a major version range lower and
upper bound to ensure that both the repository and any
online retrieval of pip result in the same major version
being deployed.

Also, the isolated fallback should not use the extra options to
ensure that it goes to the internet to find the pip wheel
to install.

Change-Id: I93a5abdbc001348075cb43a0913407d38ae89678