Provide the option to specify a list of requirements when installing pip
This patch removes the current pip pin and changes the role to provide a way
for a consumer of the role to specify any requirements they want to in a
list.

This is a more generic facility and allows the consumer to specify the
requirements in a playbook or group vars instead of being forced to
override them as they're defaulted into this role.

Change-Id: I0a78135737c40a505d77df6ed67da0ef6695bfcb