Combine pip_install and pip_lockdown roles
Since lockdown is really just a form of specific pip
configuration and since the pip_install role already does
some pip configuration, it seems logical and desirable for
these functions to take place in the same role.

This change should pave the way for a simplification of
role dependencies and the removal of explicit pip_lockdown
role usage with various playbooks that will already have a
dependency on pip_install.

Change-Id: Ia0fc276c2b501f16d4acf73bbbcad6f80804628e