Refactor pip_install to support offline deployment
Optionally support the download and installation of pip and its dependencies
via the deployment host to remove the dependence of internet connectivity on
remote containers.

When pip_offline_install is true, download get-pip.py and the pip
installation wheels to the deployment host. These wheels along with get-pip.py
are copied to the remote container and the installation is performed
offline.

Change-Id: Ic6583c54976b6fb083d9c67d055897c15bce05dc