Move required and pip packages lists into an ansible task
The "pip_packages" and "pip_required_pip_packages" have been moved
from the pip install task using the get-pip.py script because recent
versions of setuptools are failing to be installed.

Example Error: http://paste.openstack.org/show/560715/

To resolve this issue the process was broken into two parts.

Additionally when using this role outside of the gate, or in
any environment where there is not a set of pre-built wheels
available the role fails due to missing build packages.

Example Error: http://paste.openstack.org/show/560716/

To resolve this the missing packages have been added to this role.

Change-Id: I40dc3a864872dcd0ecb2d5617f652d476fd48d40
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>