Ensure that package installs can fall back
Currently the package installs only fall back if the pip
installs needed to. During upgrades from environments
where there are the right versions of pip, setuptools
and wheel in the repo, but there aren't yet the right
packages for the package install this method will fail.

This ensures that the fall back install for the packages
is decoupled from the fall back for the pip install.

Change-Id: Id855319bbab5ff6bc3b539ebedc495a5bc12afd1
Closes-Bug: #1627174