Add ability to set upper constraints file
When this role executes the installation of packages
for SNI support it should also respect upper constraints.

This patch adds the ability to set an upper constraint
file URL/path.

The tox config is also adjusted to ensure that it makes
use of upper constraints when installing requirements.

Manual backport of:
- https://review.openstack.org/388072
- https://review.openstack.org/388146
- https://review.openstack.org/388179
- https://review.openstack.org/388671

Change-Id: I2ab6f55096b560a7b443de412df168e26a3eca07