Ensure that continuation doesn't add a new line
When using a line continuation, newlines are added to the
resulting string. For the pip module's extra_args this
causes some unexpected behaviour, for example:

/usr/local/bin/pip install -U --constraint http://git.openstack.org/cgit/openstack/requirements/plain/upper-constraints.txt \n ndg-httpsclient requests

While in this particular case it ends up working, it breaks
in other cases with an error indicating that it could not
find the requirements to install.

As seen in [1] the way to ensure that there is no newline
added we adjust the way the continuation is implemented
just slightly.

[1] http://stackoverflow.com/questions/3790454/in-yaml-how-do-i-break-a-string-over-multiple-lines

Change-Id: I3d6c9dd1faeb2bf9c9c422100f54f90c1e3e5bc1