Fix same-host offline install race
When pip_offline_install is used on a deployment such as an AIO where
the host being installed to ie. "aio1" is also the deployment host,
a race condition existed where the pip wheels would be downloaded,
uploaded to the same path on the target host, and then cleaned up
from the deployment host before the installation was continued.

The result is a race condition where the wheels are removed from
the target prior to installation due to the target being local.

This changes the pip offline install cleanup to run as a handler
so the cleanup can take place after the role has completed its work
with the cached wheels.

Backport of https://review.openstack.org/#/c/389269/

Change-Id: Ibf36118d42950f0f64f83b6ffc3601c9a241289d
(cherry picked from commit d3a2adc67cac9a48f0de96e2e14f7faf1d0cf5b4)