Provide toggle for get-pip.py get_url task force option
This patch provides the ability to toggle whether the get-pip.py
script is always fetched from the source when the task is
executed.

In the integrated build, the get-pip.py script is put into the
repository, so the default of 'force: yes' is appropriate.

In role tests, however, we may wish to skip always fetching the
script to reduce errors relating to SSL validation when the task
executes a second time.

Needed-By: I8b1c570c3a0713de3e84fb4e4068e270d9e5a4dd
Change-Id: Ifb877e859a302b85c877aa867a0fc630b8765a19