Allow the usage of own repos instead of uca/rdo
If a deployer has a mirror of uca/rdo, it probably wants to use it.
Defining the URL currently works, but that process wouldn't work if
the mirror is using a different key.

This commit keeps the current behavior of using UCA with signing
packages from upstream everywhere, but with this change, you can also
define your own mirror (CentOS/Ubuntu) with your own key.

Change-Id: Icd9c3b6b7b824cb5cb29fd745777e4c6148e6481
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>