Allow the manipulation of multiple keys/repos
When doing an upgrade, or if your repo is built differently (like
per version), you probably want a facility to remove an older repo
to move to a new repo.

In order to do that, we can use a list instead of using the repo
data immediately. Some of our roles already have this behavior,
and now we make it consistent.

Change-Id: I5690ecb94c1f8a70a506d65339d7108081c138b7
(cherry picked from commit 6dd2d44b0220856389f6d2be4d42880c9145f51e)