Use an explicit version of urrlib3
If not forcing to install urllib3, ansible on the target node
will reuse the urrlib3 from site-packages, which could be older
than the one in global-requirements, because ansible doesn't
list it in its dependencies, so we don't explicit force a
certain version to be installed.

On my machine, I had an urllib3 installed with a version of
1.7.1, and the get_url/uri modules with https had SSL issues.
Moving to a more recent version urllib3 fixed it.

Change-Id: Id249d8f3aa9adab3d3b86a95bcd518c927d6a45b