Optimize pip_install
This patch optimizes the pip_install tasks to avoid lots of skipped
tasks. This role is run *many* times during the integrated gate and
adds a signficant amount of time to gate jobs.

Installing the RDO RPM still takes quite a bit of time (~ 6min) in
the integrated build, so further optimization is needed.

Change-Id: I3e127a4451a0ab47588a213e9721bc2f36b12387