Use block-rescue for pip install
Instead of using two tasks, skipping the second if the first
succeeds, this will execute the first and totally ignore the
second if the first succeeds. The second will only execute if
the first fails.

This is more efficient to execute and gets rid of the rather
horrible 'failed_when: false' on the first task.

Change-Id: I285e778fae4843f2f0c8547e2a7eb1671e240ebe