Enable idempotence test
Tasks involved in the online install of pip have been made idempotent
and the idempotence test has been enabled.

The selfcheck.json file will only be written if it does not already
exist. Blocks have been added to tasks that have a fallback if they
fail. Tasks that run commands to install or download pip packages now
check stdout for success messages to determine if they should be marked
changed.

Change-Id: I0d440efd4d95420f6f43316f5e0242b8884293c8