Remove EPEL and use RDO instead
This patch removes the epel-release package whenever the pip_install
role is applied and should eliminate package conflicts between RDO
and EPEL. The RDO package path is also adjusted to download the RDO
package from repos.fedorapeople.org (which slashes 3 HTTP redirects
out of the 5 when using rdoproject.org).

This is a breaking change for OSA deployments which are already
deployed on CentOS. Some containers may need to be rebuilt after
making this change.

Related-bug: 1670012
Change-Id: I2e69fac6368c032e237ca35839dd09dab9b0c5c0