SUSE: Add missing pre_install_zypper file for SUSE
Add missing pre_install_zypper file to SUSE which is required
by the pre_install.yml play. This also renames the SUSE variables
file to suse-42.yml to match the rest of the distributions.
Finally, the run_tests.sh, Vagrantfile and bindep.txt files are
synced with the openstack-ansible-tests repository.

Change-Id: I2fb733f5818098486e26ca7fb2ac015f02e3aa44