Pass list for package install
Ansible 2.x allows for packages to be passed as a list directly to
the `name` argument. This patch moves the EPEL packages into a list
and converts all package install tasks to pass a list directly
rather than looping over a list with with_items.

This should speed things up a bit. ;)

Change-Id: I0c185ad9d2a69f5625309d77a30536d2d373b2b5