Ensure that all repos are configured properly
It appears that yum-config-manager does not properly
configure all the repositories reliably when trying
to give it a bunch of them to do at once.

This patch sets each repo to be enabled and have the
priority set one at a time.

Change-Id: Ia56d7344572a0addfd4fd5c93832dc66f630e93d
(cherry picked from commit 843673591f32b4a86fc63257e823df5015915d60)