Rename package install task
The majority of the other roles use "Install distro packages"
as the name for the package installation task. This patch renames the
task in the pip_install role to match.

This will help when we try to profile the CI jobs to see which tasks
are taking the longest time to complete.

Change-Id: I353a4da01d7f99ccc33d9a0e4a494e9b43438da4
Signed-off-by: Major Hayden <major@mhtx.net>