Don't install any packages via pip
This patch ensures that we don't install any packages outside of a
virtualenv with pip.

CentOS has issues with python modules which are installed with pip
and are later installed with yum. Yum doesn't understand how to deal
with a package installation when some files are already there from
that package.

Closes-Bug: 1715888
Change-Id: I79e439f73e9d2790cb7a1327c1e1f4808b073e46