vars: Explicitly pull the openssl package
We are using the openssl package in various places but we never
actually pull it explicitly. Ubuntu pulls openssl as a dependency of
the lxc-templates package in the lxc_hosts role, and on openSUSE the
libopenssl-devel package requires openssl. However, on CentOS nothing
really requires the openssl executable (at least until it's needed in
the lxc_container_create role) so we may end up trying to use it
and fail if it's not already installed on the host. In the OpenStack
CI the openssl package seems to be pulled as a dependency of the
'unbound' package.

This fixes the following issue in the lxc_container_create role
when running on CentOS

"/usr/local/bin/lxc-veth-wiring: line 24: openssl: command not found"

Change-Id: Ifa6e2e476710a61a632dba8fedec786c58ae93c2