Add ability to set the default index
In this patch we allow the default index to be
changed so that we can set it to the repo server
when the pypi reverse proxy is available to be
used.

The repeated implementation of the find-links
in the install section is removed as it is
unnecessary - the global find-links option
applies there too.

Implements: blueprint python-build-install-simplification
Change-Id: I56f22f46ff849a7049ac7ae873af7a4bc526fa63
Signed-off-by: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>