Remove selfcheck.json
The file selfcheck.json has no discernable purpose,
so it is removed. The file pip-selfcheck.json is
automatically created by pip and used to track
when last it checked if it's on the latest version.

The tags implemented in the role are also corrected
to ensure that the placement of the pip.conf file
is considered a 'config' task, not an 'install'
task. The sub-tags are removed as they are unnecessary.

Change-Id: I53ad3a9c50071a34286f2351f5f245f272c07638
(cherry picked from commit 4e7e23f6e99d542468774cc4708522f330972908)