Add find-links, even when not being locked down
Using find-links when doing installs can be done
regardless of whether we're locking down the index.

In order to help transition to using the pypi cache
alongside the older repo build mechanism, we need
to be able to use a set index with find-links
pointing at the old links location.

Change-Id: I1e9782eb0fd72690a9644c7a01e8c83a4486872b
Implements: blueprint python-build-install-simplification
Signed-off-by: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>