Remove pip_lock_to_internal_repo variable
The ability to lock the pip configuration down has
now been extended so that it is possible to implement
a different index and to use find-links to achieve the
same goal.

The only replacement implementation could be to disable
the setting of an index if one wanted to *only* use
wheels, but that is unnecessary in OSA builds as we are
able to set the build to only use a specific set of
wheels through the pypiserver on the repo container.

As such, this variable has become useless and unused.
This patch therefore removes it.

Change-Id: If472c845036ab904ffc15ca7201ae68d282c1be0
Implements: blueprint python-build-install-simplification
Signed-off-by: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>