Remove the implementation of constraints in pip.conf
In https://review.openstack.org/526889 constraints were
added to pip.conf. While this may be something we want
later in the python build simplification process, it is
currently causing problems in several tests due to the
assumption that constraints are only set on the CLI.

At this stage I think it's better to leave constraints
setting to be done on the CLI as that is ubiquitous
across all pip install tasks right now.

If we switch to setting constraints in pip.conf then it
will take more co-ordination.

Implements: blueprint python-build-install-simplification
Signed-off-by: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>
Change-Id: Idf726729e725d94b70b53dc40da1ec7e7b0a00bf