Fix misspelled registered variable name in conditional
add_uca_repo -> uca_add_repo

Commit c68db4e6 refactored a task into two tasks and in the process
misspelled the variable registered in the first task within the
condition of the second task.  The undefined state of the misspelled
variable was unintentionally handled in this case because under
different circumstances the variable is expected to be undefined (ie
when the first task is skipped).

Change-Id: I73b7c5ab4de50df0b341fc3c02675e6c23b2b3da
Signed-off-by: Corey Wright <corey.wright@rackspace.com>