Only attempt package installs when packages are provided
pip has changed upstream[1] and now returns a non-zero code when
no packages are provided for install. This behavior is different
than pip<=9 in that earlier versions still returned code 0 when
an installation was called with no packages[2].

To avoid this we should only run install tasks when packages are
provided for installation.

[1] https://github.com/pypa/pip/pull/4210
[2] http://paste.openstack.org/raw/718119/

Change-Id: Iee229eb2f6f154b2f2f629c45f4ab6689d660a50