pip: Use 'upgrade-strategy' option to make PIP upgrades less aggressive
The default upgrade-strategy could change from one version of PIP to
another and the results may be unpredictable. We can control the PIP
behavior during upgrades using the 'upgrade-strategy' option and default
that to 'only-if-required' so PIP only upgrades what's really necessary.
This also ensure that getting the latest pip passes the idempotence test
now that we only upgrade only what's really necessary.

Depends-On: I998182eca9851d2cc745930fc73ca37bfefd0951
Link: https://pip.pypa.io/en/stable/user_guide/#only-if-needed-recursive-upgrade
Co-Authored-By: Jean-Philippe Evrard <jean-philippe@evrard.me>
Change-Id: I2ba89e25c0010c9a5b515a3d0c9c731b30876e74