Disable the pip version check
In our logs it is frequent that the warning to upgrade the
pip version is present, but this warning is unnecessary as
we are always using a well tested version of pip.

To reduce noise, we set disable the version check. This has
the added benefit of not adding more internet access to every
pip task than is necessary.

In the template we ensure that the var is interpreted as a
boolean (in case of user error), then force it to lower case
as that's what pip prefers in pip.conf.

We also now remove the test for the self check file, as it
will no longer be present with this option implemented.

Change-Id: Ie0427d3435270836302441b92eaff7e7980d7047
(cherry picked from commit 70b8cd5737cdf2adfe69c0cea6859118c45ae878)