Ensure that pip.conf templates variables properly
When a user provides values other than the defaults for
upgrade/pre, they may not format it in a way that pip
interprets properly. To ensure that we template the
conf file correctly, we ensure that it's interpreted
as a boolean, then make it lower case as is expected by
pip.

Change-Id: I91590665cd2d878b6b5ef92d1118610bb4a3fd84