Disable idempotence test
The idempotence of the role doesn't really work properly
with the modern get-pip.py script.

Instead of porting [1] and [2] back along with other patches
which may have merged in master but are not in the stable
branch but may be necessary, we disable the idempotence check
instead.

This way we keep the stable branch stable, but ensure
idempotence properly from Rocky onwards.

[1] https://review.openstack.org/562243
[2] https://review.openstack.org/562188

Change-Id: Iaf0ef428f9c8a845e779770fb54fe62495326c4a