defaults: Add virtualenv to default pip packages
Right now, every os_* role explicitly requires and installs the
'virtualenv' package since they all deploy virtual environments for
the services. As we move towards using distribution packages as an
alternative, we can move this dependency to the pip_install role
so we can have a central place to reference the common PIP dependencies
across the hosts.

Change-Id: I4eb0a70dc6aa4437a6aa52caf510225c49ef8b11
Implements: blueprint openstack-distribution-packages