Ensure the fallback uses the same install method
We have a test to ensure the install method of pip is from distro
or get-pip. But if an issue appeared in the first task, the
fallback method will automatically fallback to get-pip instead
of fallback to the isolated version of the pre-defined CLI.

This should ensure consistency by not using get-pip if it's not
needed.

Change-Id: I7aa4056ffb9bc2162f0960e74b6458bcbffa141f