Switch to opensuse-15
opensuse-150 nodeset is referring to openSUSE 15.0, which is still in
maintenance but openSUSE 15.1 has been released already. "opensuse-15"
is going to refer to the "latest openSUSE 15.x" build released and
working for OpenStack going forward, so add this nodeset and use
it by default going forward.

Use job openstack-ansible-functional-opensuse-15 instead of
openstack-ansible-functional-opensuse-150, thus switching nodesets.

Add job to check queue, everything in gate should be in check as well,
mark it voting.

Change-Id: Ieb48124ab26ca22720b3e1b3e8654ab89053be4d